+++ rust-sdk/core/src/constants/swap.rs
@@ -25,6 +25,6 @@ pub const MIN_SQRT_PRICE: u128 = 4295048016;
 /// The maximum sqrt price for a fusion_pool.
 pub const MAX_SQRT_PRICE: u128 = 79226673515401279992447579055;
 
-pub const MAX_CLP_TO_OLP_REWARD_RATIO: u16 = 10_000;
+pub const MAX_CLP_REWARD_RATE: u16 = 10_000;
 
 pub const PROTOCOL_FEE_RATE_MUL_VALUE: u128 = 10_000;