+++ ts-sdk/cli/src/commands/set_clp_reward_rate.ts
@@ -1,33 +1,33 @@
 import {
   fetchMaybeFusionPool,
   getFusionPoolsConfigAddress,
-  getSetClpToOlpRewardRatioInstruction,
+  getSetClpRewardRateInstruction,
 } from "@crypticdot/fusionamm-client";
 import { sendTransaction } from "@crypticdot/fusionamm-tx-sender";
 import BaseCommand, { addressArg } from "../base";
 import { rpc, signer } from "../rpc";
 import { Args } from "@oclif/core";
 
-export default class SetClpToOlpRewardRatio extends BaseCommand {
+export default class SetClpRewardRate extends BaseCommand {
   static override args = {
     pool: addressArg({
       description: "Fusion pool address",
       required: true,
     }),
-    clpToOlpRewardRatio: Args.integer({
+    clpRewardRate: Args.integer({
       description:
-        "The reward ratio of concentrated liquidity providers to limit orders' liquidity providers. The maximum value 10_000 equals to 100%",
+        "The reward rate for concentrated liquidity providers stored as basis points. The maximum value 10_000 equals to 100%",
       required: true,
       min: 0,
       max: 10_000,
     }),
   };
   static override description =
-    "Set the reward ratio of concentrated liquidity providers to limit orders' liquidity providers";
+    "Sets the reward rate for concentrated liquidity providers, generated from limit order swaps.";
   static override examples = ["<%= config.bin %> <%= command.id %> address 5000"];
 
   public async run() {
-    const { args } = await this.parse(SetClpToOlpRewardRatio);
+    const { args } = await this.parse(SetClpRewardRate);
 
     const fusionPoolsConfig = (await getFusionPoolsConfigAddress())[0];
 
@@ -39,15 +39,15 @@ export default class SetClpToOlpRewardRatio extends BaseCommand {
       throw new Error(`Fusion pool is not found at address ${args.pool}`);
     }
 
-    const ix = getSetClpToOlpRewardRatioInstruction({
+    const ix = getSetClpRewardRateInstruction({
       fusionPool: args.pool,
       fusionPoolsConfig: fusionPoolsConfig,
       feeAuthority: signer,
-      clpToOlpRewardRatio: args.clpToOlpRewardRatio,
+      clpRewardRate: args.clpRewardRate,
     });
 
     console.log("");
-    if (fusionPool.data.clpToOlpRewardRatio != args.clpToOlpRewardRatio) {
+    if (fusionPool.data.clpRewardRate != args.clpRewardRate) {
       console.log("Sending a transaction...");
       const signature = await sendTransaction(rpc, [ix], signer);
       console.log("Transaction landed:", signature);