+++ ts-sdk/cli/src/commands/set_default_clp_reward_rate.ts
@@ -1,28 +1,29 @@
 import {
   fetchMaybeFusionPoolsConfig,
   getFusionPoolsConfigAddress,
-  getSetDefaultClpToOlpRewardRatioInstruction,
+  getSetDefaultClpRewardRateInstruction,
 } from "@crypticdot/fusionamm-client";
 import { sendTransaction } from "@crypticdot/fusionamm-tx-sender";
 import BaseCommand from "../base";
 import { rpc, signer } from "../rpc";
 import { Args } from "@oclif/core";
 
-export default class SetDefaultClpToOlpRewardRatio extends BaseCommand {
+export default class SetDefaultClpRewardRate extends BaseCommand {
   static override args = {
-    defaultClpToOlpRewardRatio: Args.integer({
-      description: "The default CLP to OLP reward ratio. The maximum value of 10000 equals to 100%",
+    defaultClpRewardRate: Args.integer({
+      description:
+        "The default reward rate for concentrated liquidity providers stored as basis points. The maximum value of 10000 equals to 100%",
       required: true,
       min: 0,
       max: 10000,
     }),
   };
   static override description =
-    "Set the default reward ratio of concentrated liquidity providers to limit orders' liquidity providers";
+    "Sets the default reward rate for concentrated liquidity providers, generated from limit order swaps.";
   static override examples = ["<%= config.bin %> <%= command.id %> address 5000"];
 
   public async run() {
-    const { args } = await this.parse(SetDefaultClpToOlpRewardRatio);
+    const { args } = await this.parse(SetDefaultClpRewardRate);
 
     const fusionPoolsConfig = (await getFusionPoolsConfigAddress())[0];
 
@@ -33,14 +34,14 @@ export default class SetDefaultClpToOlpRewardRatio extends BaseCommand {
       throw new Error("FusionAMM config doesn't exists at address " + fusionPoolsConfig);
     }
 
-    const ix = getSetDefaultClpToOlpRewardRatioInstruction({
+    const ix = getSetDefaultClpRewardRateInstruction({
       fusionPoolsConfig,
       feeAuthority: signer,
-      defaultClpToOlpRewardRatio: args.defaultClpToOlpRewardRatio,
+      defaultClpRewardRate: args.defaultClpRewardRate,
     });
 
     console.log("");
-    if (config.data.defaultClpToOlpRewardRatio != args.defaultClpToOlpRewardRatio) {
+    if (config.data.defaultClpRewardRate != args.defaultClpRewardRate) {
       console.log("Sending a transaction...");
       const signature = await sendTransaction(rpc, [ix], signer);
       console.log("Transaction landed:", signature);